Create index.php
the clockframe dashboard, be gentle it's my first CSS/html/PHP project And I get the color scheme is whack but my wife likes yellow and I built all of this for her. Also the device does not have access to the internet so the PHP calls are slightly safer.